I can build against this
Hacked out of the centre of the vanguard dataplane. Needs a lot of work to polish it up, and make it coherent and generally useful (despite seeming quite sensible when in the middle of a load of other code). Abstracted over the old
managed environment, so now arbitrary environments can be managed, and the horrifying and unnecessary dependence on RoutingTables and CommandQueues is removed. Shim written in to vanguard-dataplane to re-specialise. This can also
be tidied up quite a lot once an API has been settled.

Relevant unit tests in the dataplane pass. New tests for the factored library need writing. All old tests depended on RoutingTables and CommandQueues. What the hell was I thinking.

Onwards and upwards!