Renaming functions and types in library
Renamed everything in the library with more generic names trying to capture the actual behaviour. Now everything is named in terms of "tree-threads", that is threads that have the ability to grow like trees.

Rename types

Types renamed.

Renamed files

Files are now named according to the new naming. All the code was moved in to one file to make it easier to manage (I never remember which is which anyway).

Rename functions

Functions renamed to something more generic and in line with the library.

Renaming internal variables

Renaming everything from manager type namings, to tree type namings.

Reworking comments

Comments rewritten to make more sense.